+++ R/bfactor.R
@@ -67,8 +67,8 @@
 #' \emph{Psychometrika, 75}, 581-612.
 #'
 #' Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory
-#' Package for the R Environment. \emph{Journal of Statistical Software, 48}(6),
-#' 1-29.
+#' Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.
+#' \doi{10.18637/jss.v048.i06}
 #'
 #' Gibbons, R. D., & Hedeker, D. R. (1992). Full-information Item Bi-Factor
 #' Analysis. \emph{Psychometrika, 57}, 423-436.
@@ -238,6 +238,8 @@ bfactor <- function(data, model, model2 = paste0('G = 1-', ncol(data)),
     dots <- list(...)
     if(!is.null(dots$method))
         stop('method cannot be changed for bifactor models', call.=FALSE)
+    if(!is.null(dots$formula))
+        stop('bfactor does not currently support latent regression models', call.=FALSE) #TODO
     if(missing(model)) missingMsg('model')
     if(!is.numeric(model))
         stop('model must be a numeric vector', call.=FALSE)