+++ R/randef.R
@@ -10,6 +10,13 @@
 #' @param return.draws logical; return a list containing the thinned draws of the posterior?
 #'
 #' @author Phil Chalmers \email{rphilip.chalmers@@gmail.com}
+#' @references
+#' Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory
+#' Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.
+#'
+#' Chalmers, R. P. (2015). Extended Mixed-Effects Item Response Models with the MH-RM Algorithm.
+#' \emph{Journal of Educational Measurement, 52}, 200-222. \doi{10.1111/jedm.12072}
+#' \doi{10.18637/jss.v048.i06}
 #' @keywords random effects
 #' @export randef
 #' @examples