+++ NEWS.md
@@ -1,3 +1,31 @@
+# Changes in mirt 1.25
+
+- `mdirt()` now supports latent regression covariate predictors. Associated function
+  (e.g., `fscores()`) also include the latent regression information for discrete 
+  models by default
+
+- `SIBTEST()` replaced with the asymtotic sampling distribution version of 
+   CSIBTEST described by Chalmers (accepted)
+
+- `calcNull` set to `FALSE` by default
+
+- Sandwich ACOV estimate now uses the Oakes estimate in the computations rather than the intensive 
+  Louis form (which require low-level coding of the item-level Hessian terms). Added a 
+  new `SE.type = 'sandwich.Louis'` for the original sandwich VCOV estimate in the previous version of
+  mirt
+
+- fix latent regression models with QMCEM and MCEM algorithms (reported by Seongho Bae)
+
+- `fscores()` gains a `max_theta` argument to apply upper/lower bounds to iterative searching 
+  algorithms (issue reported by Sebastian Born), and a `start` input to set the starting values
+  as well (primarily useful in mirtCAT to reduce iterations)
+
+- `alabama` package optimizer no longer used. Replaced with generic interface from `nloptr`
+  package to support numerous optimizers with greater control instead. Associated inputs
+  (e.g., `alabama_args`) replaced as well
+  
+- Export missing S4 methods for external R packages to import
+
 # Changes in mirt 1.24
 
 - MDIFF and MDISC no longer in normal ogive metric (1.702 scaling value removed)